Remove access-token handling and Logout functionality
The Backend deals with access-token handling from now on (in a cookie). Therefor, the frontend does not need to do anything except sending username and password once. As JWT has no explicit logout functionality and it is not easy to delete the token also serverside, we remove the logout functionality completely.